pyproject.toml
[project]
name = "sleeper-fantasy-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.1.1",
    "sleeper-api-wrapper>=1.1.0",
]
